Migrate e2e/e2e-test project
