+++ e2e/e2e-test/androidApp/src/main/kotlin/org/korge/e2e/MainActivity.kt
@@ -1,0 +1,10 @@
+package org.korge.e2e
+
+import korlibs.render.GameWindowCreationConfig
+import korlibs.render.KorgwActivity
+
+class MainActivity : KorgwActivity(config = GameWindowCreationConfig(msaa = 1, fullscreen = true)) {
+    override suspend fun activityMain() {
+        main()
+    }
+}
