+++ e2e/e2e-test/gradle.properties
@@ -1,10 +1,4 @@
-korgePluginVersion=7.0.0-SNAPSHOT
-#korgePluginVersion=2.6.3
 org.gradle.jvmargs=-Xmx4g
 kotlin.mpp.stability.nowarn=true
-
-kotlin.js.webpack.major.version=4
-#kotlin.test.infer.jvm.variant=false
-#org.gradle.logging.level=quiet
 
 kotlin.mpp.androidSourceSetLayoutVersion=2
