--- e2e/e2e-test/gradlew_linux
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-docker run -it -w /app \
-    -v "$(pwd)":/app \
-    -v "$HOME/.gradle_linux:/root/.gradle" \
-    -v "$HOME/.m2:/root/.m2" \
-    eclipse-temurin:21 ./gradlew --no-daemon $*
