+++ e2e/e2e-test/shared/src/commonMain/kotlin/org/korge/e2e/EmptyE2ETestCase.kt
@@ -1,0 +1,10 @@
+package org.korge.e2e
+
+import korlibs.korge.view.Container
+
+object EmptyE2ETestCase : E2ETestCase() {
+    override val pixelPerfect: Boolean = true
+
+    override suspend fun Container.run() {
+    }
+}
