+++ e2e/e2e-test/shared/src/commonMain/resources/gfx/vampire.ase
