+++ e2e/e2e-test/shared/src/commonMain/resources/gfx/vampire_slices_fixed.ase
