+++ e2e/e2e-test/shared/src/commonTest/kotlin/test.kt
@@ -1,11 +1,13 @@
-import korlibs.time.*
-import korlibs.korge.input.*
-import korlibs.korge.tests.*
-import korlibs.korge.tween.*
-import korlibs.korge.view.*
-import korlibs.image.color.*
-import korlibs.math.geom.*
-import kotlin.test.*
+import korlibs.image.color.Colors
+import korlibs.korge.input.onClick
+import korlibs.korge.tests.ViewsForTesting
+import korlibs.korge.tween.get
+import korlibs.korge.tween.tween
+import korlibs.korge.view.solidRect
+import korlibs.math.geom.Rectangle
+import korlibs.time.seconds
+import kotlin.test.Test
+import kotlin.test.assertEquals
 
 class MyTest : ViewsForTesting() {
     @Test
